fix(tidy): 修复 QmClient 首方代码静态分析问题，精简 CI 检查项
- RecordDemoInputWheelEvent() 加 const
- CBlockWordsCache 移入匿名命名空间
- FormatVoiceTransmitBlockers 参数 size_t→int 对齐 DDNet API
- 移除 modernize-use-auto（DDNet 显式转换风格 ~50 假阳性）
- 移除 modernize-use-nullptr（上游 =0 指针风格假阳性）

+++ src/game/client/components/qmclient/voice_core.cpp
@@ -2319,7 +2319,7 @@ void CRClientVoice::WorkerLoop() NO_THREAD_SAFETY_ANALYSIS
 				Preconditions.m_MicMuted = StateConfig.m_QmVoiceMicMute != 0;
 				const uint32_t TxBlockers = VoiceUtils::VoiceTransmitBlockers(Preconditions);
 				char aTxBlockers[128];
-				VoiceUtils::FormatVoiceTransmitBlockers(TxBlockers, aTxBlockers, sizeof(aTxBlockers));
+				VoiceUtils::FormatVoiceTransmitBlockers(TxBlockers, aTxBlockers, (int)sizeof(aTxBlockers));
 
 				const int TxAgeMs = m_LastTxPacketTime.load() > 0 ? (int)std::clamp((Now - m_LastTxPacketTime.load()) * 1000 / time_freq(), (int64_t)0, (int64_t)999999) : -1;
 				const int RxAgeMs = m_LastRxPacketTime.load() > 0 ? (int)std::clamp((Now - m_LastRxPacketTime.load()) * 1000 / time_freq(), (int64_t)0, (int64_t)999999) : -1;
